refactor(client-web-kit): extract error-context normalization into core/error module
