+++ packages/client-web-kit/src/core/translator/error.ts
@@ -5,70 +5,10 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import { isObject } from '@authup/kit';
 import { TranslatorTranslationNamespace } from '@authup/i18n';
 import { injectIlingo, injectLocale } from '@ilingo/vue';
 import { translateIssues } from '@ilingo/validup';
-import type { Issue } from 'validup';
-
-export type ErrorContext = {
-    code?: string;
-    data?: Record<string, any>;
-    message?: string;
-    issues?: Issue[];
-    /**
-     * The HTTP status of the transport error (a hapic `ClientError`'s
-     * `response.status`), when the error arrived over HTTP. Absent for a
-     * directly-thrown / non-HTTP error.
-     */
-    status?: number;
-};
-
-/**
- * Pull the structured `{ code, data, message, issues, status }` out of whatever
- * the caller caught. An authup server error arrives as a hapic `ClientError`
- * whose `response.data` is the serialized error body (`AuthupError.toJSON`
- * flattens `code` + `data` onto the top level and carries the validup
- * `issues`); a directly-thrown `AuthupError` or plain `Error` exposes the
- * same fields on itself. Duck-typed on purpose — no hapic import (it is not a
- * declared dependency, and its `isClientError` guard is `instanceof`-based, so
- * it would miss a non-hapic-instance error), works for both transports.
- */
-export function extractErrorContext(error: unknown): ErrorContext {
-    if (!isObject(error)) {
-        return { message: typeof error === 'string' ? error : undefined };
-    }
-
-    const self = error as Record<string, any>;
-    const { response } = self;
-    const body: Record<string, any> = isObject(response) && isObject(response.data) ?
-        response.data :
-        self;
-
-    const status = isObject(response) && typeof response.status === 'number' ?
-        response.status :
-        undefined;
-
-    // The server-side body `message` is the most specific human string; the
-    // transport error's own `message` (e.g. hapic's "Request failed") is only
-    // the last resort.
-    let message: string | undefined;
-    if (typeof body.message === 'string') {
-        message = body.message;
-    } else if (typeof self.message === 'string') {
-        message = self.message;
-    }
-
-    return {
-        code: typeof body.code === 'string' ? body.code : undefined,
-        data: body,
-        message,
-        // Only validation (`BAD_REQUEST`) errors carry issues; a coded
-        // business error serializes an empty array.
-        issues: Array.isArray(body.issues) && body.issues.length > 0 ? body.issues : undefined,
-        status,
-    };
-}
+import { extractErrorContext } from '../error';
 
 /**
  * Resolve a caught error to a localized, user-facing message.
